Search: Now supports LoadMore results - a work in progress. The status bar turns into a LoadMore button after a keyword search is run.
For each keyword variant a worker gets started. On completion each one returns the last index examined. Current strategy is to use only the index of the variant with the *most* matches returned for the next search. It is possible to "loadmore" for all variants returning matches.

On the worker side
In titleSearch mode worker can now be started with a DirEntry offset from where index traversal can resume.
For each keyword variant when worker is done, matchesfound and last examined index are now returned.
[Unrelated to load more but result presentation]
Direntry returned has a redirectedFrom prop injected if it is a redirect - This helps in presenting results and handling dups in future.
Updated Matcher fn to handle redirects

On the app side
searchDirEntries, finder can now be run with a continueFrom setting that is passed along to the worker.
This value can be obtained from the onAllWorkerCompletion callback of the last search.